split views from components
