+++ src/router/index.js
@@ -1,11 +1,11 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import Footer from '@/components/views/partial/Footer'
-import NavBar from '@/components/views/partial/NavBar'
-import Home from '@/components/views/Home'
-import NotFound from '@/components/views/NotFound'
-import SoundBoard from '@/components/views/SoundBoard'
-import LessonBoard from '@/components/views/LessonBoard'
+import Footer from '@/views/partial/Footer'
+import NavBar from '@/views/partial/NavBar'
+import Home from '@/views/Home'
+import NotFound from '@/views/NotFound'
+import SoundBoard from '@/views/SoundBoard'
+import LessonBoard from '@/views/LessonBoard'
 // import Firebase from 'firebase'
 import db from '../js/firebase-db'
